replace old access decorators
